// packages/database/src/entities/listing.entity.d.ts
import { BaseEntity } from './base.entity';
import { User } from './user.entity';
import { Agent } from './agent.entity';
import { PropertyType } from './property-type.entity';
import { AdminUnit } from './admin-unit.entity';
export declare class Listing extends BaseEntity {
  code: string;
  userId: string;
  user: User | null;
  agentId: string | null;
  agent: Agent | null;
  transactionType: 'sell' | 'rent';
  propertyTypeId: string;
  propertyType: PropertyType | null;
  title: string;
  description: string | null;
  highlights: string[];
  adminUnitCode: string;
  adminUnit: AdminUnit | null;
  address: string;
  street: string | null;
  latitude: number | null;
  longitude: number | null;
  areaLand: number | null;
  areaFloor: number | null;
  frontage: number | null;
  floors: number | null;
  bedrooms: number | null;
  bathrooms: number | null;
  direction: string | null;
  price: number;
  priceUnit: string;
  priceNegotiable: boolean;
  legalStatus: string | null;
  ownershipType: string | null;
  isMortgaged: boolean;
  amenities: string[];
  furniture: string | null;
  status: string;
  isFeatured: boolean;
  featuredUntil: Date | null;
  slug: string | null;
  views: number;
  saves: number;
  contacts: number;
  qualityScore: number | null;
  expiresAt: Date | null;
  publishedAt: Date | null;
}
//# sourceMappingURL=listing.entity.d.ts.map